+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Update `/openDevContainer` to support all dev container features when hostPath
+  and configFile are provided.
+
 ## [v1.9.2](https://github.com/coder/vscode-coder/releases/tag/v1.9.2) 2025-06-25
 
 ### Fixed